en cours

+++ Simple_Timing_Attack/KeypadEmulation.h
@@ -13,12 +13,12 @@
 #define kEYPADPIN_7 29
 class KeypadEmulation
 {
-    unsigned long buttonPushed(int row, int col, bool computeTimeThankSignal);
+    unsigned int buttonPushed(int row, int col, bool computeTimeThankSignal);
     
   public:
     KeypadEmulation();
-    unsigned long pushButton(unsigned short input, bool computeTimeThankSignal);
-    unsigned long inputSimulation(unsigned short button1, unsigned short button2, unsigned short button3, unsigned short button4);
+    unsigned int pushButton(unsigned short input, bool computeTimeThankSignal);
+    unsigned int inputSimulation(unsigned short button1, unsigned short button2, unsigned short button3, unsigned short button4);
 };
 
 #endif /* KeypadEmulation_h */
